1. Модульные тесты [#505277 #535054]

+++ src/test/java/ru/job4j/converter/ConverterTest.java
@@ -9,7 +9,7 @@ class ConverterTest {
     void whenConvert140RblThen2Euro() {
         float input = 140;
         float expected = 2;
-        float output = Converter.rubleToEuro(input);
+        double output = Converter.rubleToEuro(input);
         float value = 0.0001f;
         assertThat(output).isEqualTo(expected, withPrecision(value));
     }
@@ -19,7 +19,7 @@ class ConverterTest {
     void whenConvert180RblThen3Dollar() {
         float input = 180;
         float expected = 3;
-        float output = Converter.rubleToDollar(input);
+        double output = Converter.rubleToDollar(input);
         float value = 0.0001f;
         assertThat(output).isEqualTo(expected, withPrecision(value));
     }
